moved the file validation class to main code
